renames registry package to routing
