+++ spring-cloud-gateway-rsocket/src/main/java/org/springframework/cloud/gateway/rsocket/routing/LoadBalancerFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.rsocket.registry;
+package org.springframework.cloud.gateway.rsocket.routing;
 
 import java.util.List;
 import java.util.Random;
