+++ spring-cloud-gateway-rsocket/src/test/java/org/springframework/cloud/gateway/rsocket/autoconfigure/GatewayRSocketAutoConfigurationTests.java
@@ -28,9 +28,9 @@ import org.springframework.boot.rsocket.server.RSocketServerBootstrap;
 import org.springframework.boot.rsocket.server.RSocketServerFactory;
 import org.springframework.boot.test.context.runner.ReactiveWebApplicationContextRunner;
 import org.springframework.cloud.gateway.rsocket.core.GatewayServerRSocketFactoryCustomizer;
-import org.springframework.cloud.gateway.rsocket.registry.RegistrySocketAcceptorFilter;
-import org.springframework.cloud.gateway.rsocket.registry.RoutingTable;
-import org.springframework.cloud.gateway.rsocket.registry.RoutingTableRoutes;
+import org.springframework.cloud.gateway.rsocket.routing.RoutingTable;
+import org.springframework.cloud.gateway.rsocket.routing.RoutingTableRoutes;
+import org.springframework.cloud.gateway.rsocket.routing.RoutingTableSocketAcceptorFilter;
 import org.springframework.cloud.gateway.rsocket.socketacceptor.GatewaySocketAcceptor;
 import org.springframework.cloud.gateway.rsocket.socketacceptor.SocketAcceptorPredicate;
 import org.springframework.cloud.gateway.rsocket.socketacceptor.SocketAcceptorPredicateFilter;
@@ -54,7 +54,7 @@ public class GatewayRSocketAutoConfigurationTests {
 								MetricsAutoConfiguration.class))
 				.run(context -> assertThat(context).hasSingleBean(RoutingTable.class)
 						.hasSingleBean(RoutingTableRoutes.class)
-						.hasSingleBean(RegistrySocketAcceptorFilter.class)
+						.hasSingleBean(RoutingTableSocketAcceptorFilter.class)
 						.hasSingleBean(GatewayServerRSocketFactoryCustomizer.class)
 						.hasSingleBean(GatewayRSocketProperties.class)
 						.hasSingleBean(GatewaySocketAcceptor.class)
