+++ spring-cloud-gateway-rsocket/src/test/java/org/springframework/cloud/gateway/rsocket/routing/RoutingTableRoutesTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.rsocket.registry;
+package org.springframework.cloud.gateway.rsocket.routing;
 
 import java.util.HashMap;
 import java.util.HashSet;
